requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
APScheduler==3.0.0
attrs==16.3.0
awscli==1.11.45
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.3
billiard==3.5.0.2
botocore==1.5.8
celery==4.0.2
certifi==2017.1.23
cffi==1.9.1
colorama==0.3.7
constantly==15.1.0
contextlib2==0.5.4
cryptography==1.7.1
cssselect==1.0.0
cycler==0.10.0
diff-match-patch==20121119
dj-database-url==0.4.1
Django==1.10.4
django-celery-beat==1.0.1
django-el-pagination==3.0.1
django-import-export==0.5.1
django-querycount==0.5.0
docutils==0.13.1
enum34==1.1.6
factory-boy==2.8.1
fake-useragent==0.1.7
Faker==0.7.11
feedfinder2==0.0.1
feedparser==5.1.3
flower==0.9.1
funcsigs==1.0.2
futures==3.0.5
geograpy==0.3.7
gunicorn==19.6.0
heroku3==3.2.2
idna==2.2
incremental==16.10.1
ipaddress==1.0.18
jellyfish==0.5.6
jieba==0.35
jmespath==0.9.1
kombu==4.0.2
lxml==3.7.3
matplotlib==1.5.3
mock==2.0.0
newrelic==2.82.1.63
newspaper==0.0.9.8
nltk==3.2.1
nose2==0.6.5
numpy==1.11.2
packaging==16.8
parsel==1.1.0
pbr==1.8.0
Pillow==2.5.1
postgres==2.2.1
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycountry==16.11.27.1
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.11
queuelib==1.4.2
raven==6.0.0
redis==2.10.5
requests==2.3.0
rsa==3.4.2
s3transfer==0.1.10
Scrapy==1.3.0
scrapy-fake-useragent==1.0.2
service-identity==16.0.0
simplejson==3.3.1
six==1.9.0
sqlparse==0.2.2
stevedore==1.8.0
tablib==0.11.3
tldextract==1.5.1
tornado==4.2
Twisted==16.6.0
tzlocal==1.3
vine==1.1.3
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
w3lib==1.16.0
whitenoise==3.2.2
zope.interface==4.3.3
